refactor(bin): extract shared BundlerInput assembly and plugin-setup phase
Eliminates two verbatim duplications between zfb build and zfb dev:

(a) Config→BundlerInput assembly (~200 lines, ~25 fields) extracted into
    crates/zfb/src/commands/bundler_input.rs::assemble_bundler_input().
    Dev/build differences are explicit parameters:
      • bundle_mode: BundleMode::Production vs BundleMode::Development
      • css_fail_mode: CssModuleFailMode::HardFail vs WarnAndEmpty

(b) Plugin lifecycle setup (spawn → setup → virtual-module prefetch loop
    → alias/virtual-module derivation → translate_setup_registries_to_hooks)
    extracted into plugins.rs::run_plugin_setup(), returning PluginSetupResult.
    Dev/build difference is explicit: SetupCommand::Build vs SetupCommand::Dev.
    Steps that follow (preBuild out_dir, islands config, dev middleware,
    injected routes) remain command-local.

Closes #906.
